# Conflicts:
#	data/assets/util/asset_helper.asset
#	data/assets/util/webgui.asset

+++ src/navigation/waypoint.cpp
@@ -2,7 +2,7 @@
  *                                                                                       *
  * OpenSpace                                                                             *
  *                                                                                       *
- * Copyright (c) 2014-2021                                                               *
+ * Copyright (c) 2014-2022                                                               *
  *                                                                                       *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this  *
  * software and associated documentation files (the "Software"), to deal in the Software *
@@ -90,7 +90,7 @@ double Waypoint::validBoundingSphere() const {
 }
 
 double Waypoint::findValidBoundingSphere(const SceneGraphNode* node) {
-    double bs = static_cast<double>(node->boundingSphere());
+    double bs = node->boundingSphere();
 
     const double minValidBoundingSphere =
         global::navigationHandler->pathNavigator().minValidBoundingSphere();
@@ -101,7 +101,7 @@ double Waypoint::findValidBoundingSphere(const SceneGraphNode* node) {
         // Alsp. the chance to find a bounding sphere that represents the visual size of
         // the target well is higher for these nodes
         for (SceneGraphNode* child : node->children()) {
-            bs = static_cast<double>(child->boundingSphere());
+            bs = child->boundingSphere();
             if (bs > minValidBoundingSphere) {
                 LWARNING(fmt::format(
                     "The scene graph node '{}' has no, or a very small, bounding sphere. "
